Everything compiling, running, but weird image

+++ src/objects/camera.rs
@@ -28,8 +28,8 @@ impl Camera {
 
     // pixel coordinates to image plane transformation
     pub fn pixel_to_img(&self, x: usize, y: usize) -> Vector3 {
-        let alpha = ((x as i32 / self.width) as f32) - 0.5; // alpha and beta are normalisation factors. Here I have assumed FOV is 45 degrees and that the screen varies from 0 to 1.
-        let beta = ((y as i32/ self.height) as f32) - 0.5;
+        let alpha = (x as f32 / self.width as f32) - 0.5; // alpha and beta are normalisation factors. Here I have assumed FOV is 45 degrees and that the screen varies from 0 to 1.
+        let beta = (y as f32 / self.height as f32) - 0.5;
     
         let fovX = 3.141593 / 2.0;
         let fovY = ((self.height / self.width) as f32) * fovX;
